// packages/plugin-kit/src/ipc-worker/generated/sensync2_plugin_kit/ipc_worker/numeric_array.ts
export const SampleFormat = {
  SAMPLE_FORMAT_UNSPECIFIED: 0,
  SAMPLE_FORMAT_F32: 1,
  SAMPLE_FORMAT_F64: 2,
  SAMPLE_FORMAT_I16: 3,
  UNRECOGNIZED: -1,
} as const;

export type SampleFormat = typeof SampleFormat[keyof typeof SampleFormat];
// ...
  export type SAMPLE_FORMAT_UNSPECIFIED = typeof SampleFormat.SAMPLE_FORMAT_UNSPECIFIED;
  export type SAMPLE_FORMAT_F32 = typeof SampleFormat.SAMPLE_FORMAT_F32;
  export type SAMPLE_FORMAT_F64 = typeof SampleFormat.SAMPLE_FORMAT_F64;
  export type SAMPLE_FORMAT_I16 = typeof SampleFormat.SAMPLE_FORMAT_I16;
  export type UNRECOGNIZED = typeof SampleFormat.UNRECOGNIZED;
// ...
export interface NumericArray {
  sampleFormat: SampleFormat;
  length: number;
  data: Uint8Array;
}

function createBaseNumericArray(): NumericArray {
  return { sampleFormat: 0, length: 0, data: new Uint8Array(0) };
}

export const NumericArray: MessageFns<NumericArray> = {
  encode(message: NumericArray, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.sampleFormat !== 0) {
      writer.uint32(8).int32(message.sampleFormat);
    }
    if (message.length !== 0) {
      writer.uint32(16).uint32(message.length);
    }
    if (message.data.length !== 0) {
      writer.uint32(26).bytes(message.data);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): NumericArray {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseNumericArray();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.sampleFormat = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.length = reader.uint32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.data = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): NumericArray {
    return {
      sampleFormat: isSet(object.sampleFormat)
        ? sampleFormatFromJSON(object.sampleFormat)
        : isSet(object.sample_format)
        ? sampleFormatFromJSON(object.sample_format)
        : 0,
      length: isSet(object.length) ? globalThis.Number(object.length) : 0,
      data: isSet(object.data) ? bytesFromBase64(object.data) : new Uint8Array(0),
    };
  },

  toJSON(message: NumericArray): unknown {
    const obj: any = {};
    if (message.sampleFormat !== 0) {
      obj.sampleFormat = sampleFormatToJSON(message.sampleFormat);
    }
    if (message.length !== 0) {
      obj.length = Math.round(message.length);
    }
    if (message.data.length !== 0) {
      obj.data = base64FromBytes(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<NumericArray>, I>>(base?: I): NumericArray {
    return NumericArray.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<NumericArray>, I>>(object: I): NumericArray {
    const message = createBaseNumericArray();
    message.sampleFormat = object.sampleFormat ?? 0;
    message.length = object.length ?? 0;
    message.data = object.data ?? new Uint8Array(0);
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}